:truck: Rename and order Settings module files and folders for more clarity
