+++ src/components/screens/languageSelection/index.js
@@ -6,7 +6,7 @@ import { colors } from 'constants/styleGuide';
 import { P } from 'components/shared/toolBox/typography';
 import Icon from 'components/shared/toolBox/icon';
 import withTheme from 'components/shared/withTheme';
-import { settingsUpdated as settingsUpdatedAction } from 'modules/Settings/actions';
+import { settingsUpdated as settingsUpdatedAction } from 'modules/Settings/store/actions';
 import getStyles from './styles';
 
 @connect(
