+++ src/modules/Accounts/components/AccountItem/index.js
@@ -36,7 +36,7 @@ export default function AccountItem({
           color: colors.dark.blueGray,
           icon: () => <RefreshSvg />,
           testID: 'backup-account',
-          onPress: () => navigation.navigate('PassphraseBackup', { account }),
+          onPress: () => navigation.navigate('BackupPassphrase', { account }),
         },
       ]}
       rightActions={[
