+++ src/modules/Settings/components/BackupPassphrase/BackupPassphrase.styles.js
