+++ src/modules/Settings/components/BackupPassphrase/components/CopyPassphrase.js
@@ -1,23 +1,30 @@
 import React, { useState } from 'react';
 import { View, ScrollView, SafeAreaView } from 'react-native';
-import { translate } from 'react-i18next';
+import { useNavigation } from '@react-navigation/native';
+import i18next from 'i18next';
 import QRCode from 'react-native-qrcode-svg';
+
 import { themes, colors } from 'constants/styleGuide';
+import { useTheme } from 'contexts/ThemeContext';
 import { deviceWidth, deviceHeight, SCREEN_HEIGHTS } from 'utilities/device';
 import { PrimaryButton } from 'components/shared/toolBox/button';
-import PassphraseCopy from 'components/shared/passphraseCopy';
-import withTheme from 'components/shared/withTheme';
+import CopyPassphraseToClipboard from 'components/shared/CopyPassphraseToClipboard/CopyPassphraseToClipboard';
 import { P, A } from 'components/shared/toolBox/typography';
 import HeaderBackButton from 'components/navigation/headerBackButton';
-import { useNavigation } from '@react-navigation/native';
-import getStyles from './styles';
+
+import { getCopyPassphraseStyles } from './CopyPassphrase.styles';
 
 const isSmallScreen = deviceHeight() < SCREEN_HEIGHTS.MD;
 const qrCodeSize = deviceWidth() * (isSmallScreen ? 0.64 : 0.72);
 
-const PassphraseBackup = ({ styles, t, theme, sharedData: data, nextStep }) => {
+export default function CopyPassphrase({ sharedData: data, nextStep }) {
   const [passphraseRevealed, setPassphraseReveal] = useState(false);
+
   const navigation = useNavigation();
+
+  const { styles, theme } = useTheme({
+    styles: getCopyPassphraseStyles(),
+  });
 
   const toggleQRCode = () => {
     setPassphraseReveal(!passphraseRevealed);
@@ -28,16 +35,23 @@ const PassphraseBackup = ({ styles, t, theme, sharedData: data, nextStep }) => {
   };
 
   return (
-    <SafeAreaView style={[styles.wrapper, styles.theme.wrapper]}>
-      <HeaderBackButton title={'settings.backupPhrase.title'} onPress={navigation.goBack} />
-      <ScrollView style={styles.container}>
-        <PassphraseCopy passphrase={data.recoveryPhrase} />
+    <SafeAreaView style={[styles.container, styles.theme.container]}>
+      <HeaderBackButton
+        title={i18next.t('settings.backupPhrase.title')}
+        onPress={navigation.goBack}
+      />
+
+      <ScrollView style={styles.body}>
+        <CopyPassphraseToClipboard passphrase={data.recoveryPhrase} />
+
         <View style={styles.row}>
-          <P style={[styles.QRText, styles.theme.text]}>{t('Private use only')}</P>
+          <P style={[styles.qrText, styles.theme.text]}>{i18next.t('Private use only')}</P>
+
           <A style={styles.button} onPress={toggleQRCode}>
-            {passphraseRevealed ? t('Hide QR code') : t('Show QR code')}
+            {passphraseRevealed ? i18next.t('Hide QR code') : i18next.t('Show QR code')}
           </A>
         </View>
+
         {passphraseRevealed && (
           <View style={styles.qrCodeContainer}>
             <QRCode
@@ -51,11 +65,12 @@ const PassphraseBackup = ({ styles, t, theme, sharedData: data, nextStep }) => {
           </View>
         )}
       </ScrollView>
-      <View style={styles.container}>
-        <PrimaryButton noTheme title={t('settings.backupPhrase.continue')} onPress={nextScreen} />
+
+      <View style={styles.footer}>
+        <PrimaryButton onClick={nextScreen} noTheme>
+          {i18next.t('settings.backupPhrase.continue')}
+        </PrimaryButton>
       </View>
     </SafeAreaView>
   );
-};
-
-export default withTheme(translate()(PassphraseBackup), getStyles());
+}
