+++ src/modules/Settings/store/actionTypes/index.js
