+++ src/modules/Settings/store/actions/settings.test.js
