+++ src/modules/Settings/store/reducer/settings.test.js
