+++ src/store/reducers/index.js
@@ -1,5 +1,5 @@
 export { default as account } from 'modules/Accounts/store/reducer/account';
-export { default as settings } from 'modules/Settings/reducer/settings';
+export { default as settings } from 'modules/Settings/store/reducer/settings';
 export { default as bookmark } from 'modules/Bookmark/store/reducer';
 export { default as loading } from './loading';
 export { default as service } from './service';
